+++ algonaut_algod/src/models/asset_holding.rs
@@ -10,16 +10,14 @@
 
 /// AssetHolding : Describes an asset held by an account.  Definition: data/basics/userBalance.go : AssetHolding
 
-
-
 #[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
 pub struct AssetHolding {
     /// \\[a\\] number of units held.
     #[serde(rename = "amount")]
-    pub amount: i32,
+    pub amount: u64,
     /// Asset ID of the holding.
     #[serde(rename = "asset-id")]
-    pub asset_id: i32,
+    pub asset_id: u64,
     /// \\[f\\] whether or not the holding is frozen.
     #[serde(rename = "is-frozen")]
     pub is_frozen: bool,
@@ -27,7 +25,7 @@ pub struct AssetHolding {
 
 impl AssetHolding {
     /// Describes an asset held by an account.  Definition: data/basics/userBalance.go : AssetHolding
-    pub fn new(amount: i32, asset_id: i32, is_frozen: bool) -> AssetHolding {
+    pub fn new(amount: u64, asset_id: u64, is_frozen: bool) -> AssetHolding {
         AssetHolding {
             amount,
             asset_id,
@@ -35,5 +33,3 @@ impl AssetHolding {
         }
     }
 }
-
-
