+++ algonaut_algod/src/models/transaction_params_200_response.rs
@@ -10,8 +10,6 @@
 
 /// TransactionParams200Response : TransactionParams contains the parameters that help a client construct a new transaction.
 
-
-
 #[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
 pub struct TransactionParams200Response {
     /// ConsensusVersion indicates the consensus protocol version as of LastRound.
@@ -19,7 +17,7 @@ pub struct TransactionParams200Response {
     pub consensus_version: String,
     /// Fee is the suggested transaction fee Fee is in units of micro-Algos per byte. Fee may fall to zero but transactions must still have a fee of at least MinTxnFee for the current network protocol.
     #[serde(rename = "fee")]
-    pub fee: i32,
+    pub fee: u64,
     /// GenesisHash is the hash of the genesis block.
     #[serde(rename = "genesis-hash")]
     pub genesis_hash: String,
@@ -28,15 +26,22 @@ pub struct TransactionParams200Response {
     pub genesis_id: String,
     /// LastRound indicates the last round seen
     #[serde(rename = "last-round")]
-    pub last_round: i32,
+    pub last_round: u64,
     /// The minimum transaction fee (not per byte) required for the txn to validate for the current network protocol.
     #[serde(rename = "min-fee")]
-    pub min_fee: i32,
+    pub min_fee: u64,
 }
 
 impl TransactionParams200Response {
     /// TransactionParams contains the parameters that help a client construct a new transaction.
-    pub fn new(consensus_version: String, fee: i32, genesis_hash: String, genesis_id: String, last_round: i32, min_fee: i32) -> TransactionParams200Response {
+    pub fn new(
+        consensus_version: String,
+        fee: u64,
+        genesis_hash: String,
+        genesis_id: String,
+        last_round: u64,
+        min_fee: u64,
+    ) -> TransactionParams200Response {
         TransactionParams200Response {
             consensus_version,
             fee,
@@ -47,5 +52,3 @@ impl TransactionParams200Response {
         }
     }
 }
-
-
